requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
ultralytics==8.0.196
opencv-python==4.8.1.78
pillow==10.0.1
numpy==1.24.3
python-dotenv==1.0.0
gunicorn==21.2.0
tkinter
